test: add comprehensive test coverage for CLI and core modules
Add extensive test coverage for CLI commands, system orchestrator, and base service functionality. Tests cover error handling, signal management, exception hooks, health checks, and various CLI command scenarios including success, partial success, and error cases.

+++ src/ai_unit_test/core/implementations/llm/openai_connector.py
@@ -219,7 +219,7 @@ class OpenAIConnector(LLMConnector[OpenAIConnectorConfig]):
             )
 
             response_time_ms = int((time.time() - start_time) * 1000)
-            embeddings = np.array([item.embedding for item in response.data])
+            embeddings = np.array([item.embedding for item in response.data], dtype=np.float32)
 
             if request.normalize:
                 from sklearn.preprocessing import normalize  # type: ignore[import-untyped]
